[FLINK-10517][rest] Extract DocumentingDispatcherRestEndpoint
